Exp4/generation_s2/Imageio/_m3_requirements_raw.txt
<requirements>
numpy
Pillow
</requirements>